Move 5 files from functional to integration
