+++ test/integration/auth/scram_sha_1.test.js
@@ -1,7 +1,7 @@
 'use strict';
-var test = require('./shared').assert;
-var setupDatabase = require('./shared').setupDatabase;
-const { MongoClient } = require('../../src');
+var test = require('../../functional/shared').assert;
+var setupDatabase = require('../../functional/shared').setupDatabase;
+const { MongoClient } = require('../../../src');
 const { expect } = require('chai');
 
 describe('SCRAM', function () {
